requirements.txt
Django==4.0.1
gunicorn==20.1.0
django-heroku==0.3.1
django-storages==1.12.3
google-auth==2.11.0
google-cloud-storage==2.5.0